Breaking change: prepared queries now take the values in an array, instead of using square brackets to set their values:
query.execute(["bob", "baz"]);

instead of

query[0] = "bob";
query[1] = "baz";
query.execute();

This is because the old way wasn't safe. In this example of the old way,
the query is actually run with a parameter of "thing9" every time.

for (var i = 0; i < 10; i++) {
  query[0] = "thing$i";
  query.execute();
}

Now, it works correctly:

for (var i = 0; i < 10; i++) {
  query.execute(["thing$i"]);
}