Fix executeMulti.
When I rewrote results to always be a stream, I missed that executeMulti can't stream the results of the query as it needs to run another query before the results are accessed. I fixed it by reading the stream of rows into a list, and returning that as a stream.