Fix bug when .query() fails if first returned element is blank string
We are considering that if the first byte in the response equals to '00',
this is an indicator of 'okPacket', but it appears MySQL also sends '00'
if the first column in the record is a blank string.

To fix that, I added an additional check that we are not in the state
when we are expecting row results, when we considering that an
'okPacket'.